Functions should return values at the End and in some of your other problems method chaining can be done [Without breaking readability]
Make sure to use const and let while using javascript. A problem I want to notify is while we use functions without the return keyword undefined is being returned from the function (odd()) implicitly. 

PS: In other languages, a function that returns undefined is referred to as a "procedure." 

Reference from StackOverflow: https://stackoverflow.com/questions/53197907/is-it-a-bad-practise-to-not-return-anything-from-js-function